+++ src/Command/DeployTasksShowCommand.php
@@ -12,6 +12,8 @@ use Soviann\DeployTasksBundle\Storage\TaskExecution;
 use Soviann\DeployTasksBundle\Storage\TaskStorageInterface;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Completion\CompletionInput;
+use Symfony\Component\Console\Completion\CompletionSuggestions;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -32,6 +34,13 @@ final class DeployTasksShowCommand extends Command
         private readonly TaskDescriptionResolver $descriptionResolver,
     ) {
         parent::__construct();
+    }
+
+    public function complete(CompletionInput $input, CompletionSuggestions $suggestions): void
+    {
+        if ($input->mustSuggestArgumentValuesFor('id')) {
+            $suggestions->suggestValues(\array_keys($this->registry->allRegistered()));
+        }
     }
 
     protected function configure(): void